Add more log to GitHub callback endpoint
In our GitHub callback endpoint, we conduct multiple checks and return
relevant error messages to the user. However, if the user doesn't share
the error message they received, it becomes challenging to troubleshoot
the installation issue. To address this, this PR introduces logs for
each check, enabling us to effortlessly track which check failed.