Remove downloading cache proxy and forked runner
Since proxy code will be working via a service installed
on the runner image, don't need to download and run the code directly.

We've also started to use forked version of the runner as a part of
the runner images, so no need to explicitly install it anymore too.

Updating CUSTOM_ACTIONS_CACHE_URL is still necessary to send cache api
requests to local proxy.

Once we'll make sure to use transparent cache for all users, that
part will be moved out of project flag check later.