Add password requirements for user accounts
Previously, we used the default password requirements that comes with rodauth,
which are 6 characters minimum and no requirement on the type of characters.

The defaults are too permissive. With this commit we are increasing minimum
length to 8 and also require passwords to contain at least one lowercase, one
uppercase and one digit.

Finally, we also set a maximum length as well, this closes a low risk attack
vector (DoS with long password). The length 72 is used becase it seems bcrypt
only uses first 72 bytes of password while creating the hash. Source:
https://github.com/jeremyevans/rodauth/blob/master/doc/login_password_requirements_base.rdoc