Send mail to partners when new database is created
As our partnership discussion, partners will be responsible for the customer
journey and success. For effective communication, partners would need contact
information of the customers. This commit adds logic for sending e-mail to the
partners when a database managed by them is created.

In the UI, we will ask consent for sharing contact information.