Add managed access policies
One of the biggest issues in our user invitation flow is that invited
users are not granted any permissions by default. The individual who
invites the new user must manually assign permissions using our JSON
policy editor. Often, users are not aware they need to do this.
Moreover, writing a JSON policy can be challenging for newcomers.

To streamline the user invitation process, we introduced managed
policies for new users. These policies overlap with popular roles in
other systems. Assigning roles to users is common practice in many
applications.

The ACLs of these policies are hidden from the user, who only sees the
policy name. The controlplane manages the policy content.

When the first user is added to this policy, we create a new access
policy for the project. If the access policy already exists, we simply
overwrite its body.