Allow to select a policy for the users of project
I've added a selection box to each row in the user list. Users can
select a policy for their project and save it.

To identify a user's current policy, we retrieve the managed access
policies of the project and get the accounts for each policy using the
'subjects' property of the policy via access tag.

To set managed policies for users, we get the policies for all users of
the project in a single HTML form request, formatted as
`user_policies[<ACCOUNT_UBID>] = <POLICY_NAME>`. We simply replace all
policy content with these users.

We've decided to revise the user management page's design, resulting in
a few more changes than a single picker would entail.