Allow to select a policy while inviting a user
If the user invited to the project, but doesn't have an account at
Ubicloud, we persist the invitation until they create one.

With this PR, the inviter can select a policy while inviting an user.
Once the invited user creates an account, we append them to the existing
policy's subjects list instead of replacing the entire list.

Additionally, the policy for the invited user can be updated.