New boot image types for AI inferencing
This commit introduces new boot image types related to AI inferencing.
The new types are: AI base and AI model boot images. The AI base boot
image is a base image that contains the necessary software to run
inferencing. The AI model boot image is a boot image that contains the
AI model to be run on the AI base boot image.