Create new project with managed admin policy instead of custom policy
When we create a new project, we also create an access policy granting
all permissions to the current user. With the introduction of managed
policies, we can now apply the managed admin policy, eliminating the
need to create a custom policy with full permissions.

The current user will be admin for the newly created project.

Most changes are within the test, as we expect the project have a custom
admin policy.