Make advanced access policy optional
Since we grant initial permissions using a managed policy, not an
advanced access policy, the project may not have an advanced policy.

Therefore, if the project lacks an advanced policy, we first create a
new one.

Additionally, our backend supports multiple advanced policies, but for
simplicity, we only display one in the UI. I've simplified the routes to
use just one advanced policy.