Send a link to create an account if the invited user doesn't already have
We've sent the project dashboard link to the invited user. However, as
login is required to access the project page, a warning prompts them to
log in.

This process can be confusing for users without an account. To clarify,
I changed it to send a link for account creation in the invitation if it
doesn't have an account.

Also, I refactored the if-else block for readability, separating failure
and success cases into two distinct blocks.