Self-signed load balancer certificates
Load balancer certificates currently require a configured DNS zone.
In development environments we typically don't have that setup. That
makes it hard to provision a functional load balancer in development
environments. In this commit we create a self-signed load balancer
certificate in development environments that don't have a dns zone setup.