Use local ip instead of localhost for cache endpoints
To access the local cache server on port 51123, localhost was
used. It was ok to use that to access the server via processes
running on the VM itself. Though, processes running in a container
can not access the cache server with localhost. So, using local
ip address instead to allow processes from both VM and container
to access cache server.