Divide Dependabot version updates for production and development
As our number of dependencies grows, it results in large PRs that are
difficult to review. The recent aws-sdk-s3 failures highlight the need
for us to be more cautious with these updates.

Dependabot has a feature that allows us to group updates by type,
enabling us to update development dependencies separately from
production dependencies.