Clear explicit dns resolution for runners
Since we were not using dnsmasq to resolve dns names on github runners,
we've added an explicit configuration to resolved configuration. So
runner VMs can resolve names directly using the information passed
on the configuration file. As we've decided to use dnsmasq for
resolution, removing explicit configuration.

Note that, this commit will be removed once the explicit configuration
setting is removed from the runner image.