Assign private IPv4 address to Dnsmasq
We had reserved the first 4 addresses from the private subnet for our
own usage. They sit still and now we need one of them. Ubicloud VMs are
totally fine using IPv6 for dns queries, however, the container
landscape is not like that... The docker containers running on these VMs
prefer (have to) use IPv4. This is why our dnsmasq based dns solution
tend to break for containers. Here, in this commit, we are assigning the
second address from the private subnet to dns. AWS does the same thing.
More information available here:
https://docs.aws.amazon.com/vpc/latest/userguide/subnet-sizing.html