Fix the condition check during installation destroy
We ensure that the installation has no repositories or runners before
destroy it.

I attempted to handle this check on the query side and used `.exists`
with the dataset. However, it doesn't work as expected. Instead of
returning a boolean, it returns `Sequel::SQL::PlaceholderLiteralString`.
The `if` statement always evaluates as true because it's an object. I
replaced it with `empty?`.