Add Load Balancer custom hostname support
Before this commit, we would use the name of the lb, the first 5 chars
of the subnet ubid and .lb.ubicloud.com as the hostname of the
load balancer. With this commit, we start allowing first level
integrating customers (such as AI service) to specify their own dns zone
and the hostname. This way, our load balancers can have nicer names for
internal services.