Only show Lantern PostgreSQL as a provisioning option to selected projects
This is a soft feature flag. Technically, it is still possible for people to
send requests to backend but I think the probability of people discovering this
is low and even if they discover it, they are welcome to use it. The primary
reason for feature flag is giving us some time until we finalize the copy in
the UI and the pricing.