Use strings instead of symbols in Postgres configs map
These are used to generate JSON config hash, which already convers symbols to
strings. Additionally, config names with . in them such as `cron.database_name`
looks ugly when used as symbol. Basically, we need to encapsulate them with
quotation marks anyway. So, we decided to use strings directly instead of
symbols.