Do not allow to remove all admins from project
It's possible for a user to remove all of their permissions, leaving the
project inaccessible. A user might also mistakenly strip away all their
own permissions, again resulting in an inaccessible project. This new
condition check prevents the updating of management policies if there's
no remaining admin in the project.