Add active? helper to the project
We need to determine whether a project is active or not to allow certain
actions, like installing our GitHub App. This is particularly important
when we're investigating the project for any fraudulent activity. I've
added two basic conditions to assess a project's activity status, and we
can improve this in the future.

At present, if any account within the project is suspended, we consider
the project inactive. We also soft delete projects, preserving their
billing information temporarily to finalize the billing. If a project is
'soft deleted', we classify it as inactive.