Do not poll jobs if the project is inactive
We have a job poller that checks the repository for queued jobs via
GitHub API, ensuring no webhook events are missed and that a runner is
provisioned for each one.

However, when we suspend an account, we intentionally avoid provisioning
runners. Despite this, the poller might misinterpret this as the
repository lacking runners and provision new ones.

Ideally, we should destroy GithubInstallation when we suspend an account.
If for some reason this hasn't been done, we shouldn't poll the jobs.