Do not register runner for inactive projects
In the last wave of fraud, we prevented new runner registrations for
fraudulent accounts using our quota system. However, we can simplify
this. If the project is inactive, we can terminate the runner prog. A
project's activity status is determined by suspended accounts and soft
project deletions. As the job poller also checks project activity, we
can exit from this prog instead of taking a long nap. The poller will
not a create new runners.