Return related error message when the cache entry already exists
`GithubCacheEntry` is unique by `[:repository_id, :scope, :key,
:version]`. We check for its existence and return a smoother error
message when the cache entry already exists.

If a customer tries to create a cache entry that already exists at the
same time, a race condition could cause two threads to attempt to create
the same cache entry, resulting in an HTTP 500 error. We see a few
examples of this every day.

Catching the exception for an already existing cache entry makes it more
robust.

We already have a test for the already existing case.