Do not autoload in production or when running with NO_AUTOLOAD
Previously, Ubicloud setup autoloads for all constants, then referenced
all constants to force the loads. This was necessary previously as the
loaded files referenced constants without requiring them, relying on the
autoloads to work.  Recent commits have fixed that, and allow all files
to work when required directly.

This removes the automatic generation of empty modules, as all modules
needed are now defined directly in the loader.  Since the files are
required and not autoloaded, this removes the tracking of autoloaded
constants, and the forced referencing of them in production.