Add connected subnet entity management functions
Here we are adding the methods to be able to find connected_subnets,
dis/connect 2 subnets, find out all the related nics. The
find_all_connected_nics function is interesting and needs further
explanation.

find_all_connected_nics simply find all the nics that are kind of
related to the subnet we are in. For example, let's assume we have three
subnets that are connected to each other in this way A<->B<->C. Let's
also assume that each of these subnets have 1 nic. To be able to
implement connected subnets, we need to implement IpsecTunnels in
between nicA<->nicB AND nicB<->nicC. Furthermore, to be able to
implement tunnel destroys in rekey nic tunnel, we remove old states to
avoid leaving ip xfrm state objects dangling. Therefore, in a graph, if
we are performing rekeying for a subset of tunnels, we need to rekey all
the active tunnels. Otherwise, they will get dropped. For that reason,
we find all the connected nics by performing depth first search.

Let's also explain what we do to connect_subnet and disconnect_subnet.
We first create a ConnectedSubnet entity, and then, we create tunnels
from each nic in the existing subnet to every other nic in the
connected_subnet. Therefore, if we have 3 nics in the current subnet and
2 nics in the other, we create 6 tunnels in total. We destroy these
tunnels at the tume of disconnect_subnet. However, we try to be careful
not to destroy tunnels that are in between the nics in the same subnet.