Run finalize_associations on each Sequel::Model class when testing
This is something Sequel::Model.freeze_descendants does, which
is what is run in production.  It would be best to be able to
use Sequel::Model.freeze_descendants when testing, but due to
the extensive use of rspec mocking, that isn't currently
possible.  This makes the test environment closer to the
production environment, and it speeds up the tests.

While here, change freeze_descendents to freeze_descendants. Both
are currently supported, but descendents is a typo, and I plan to
remove freeze_descendents in Sequel 6 (see https://github.com/jeremyevans/sequel/commit/185efe54047a6c912555651523726f9929451589)