Only run E2E GitHub Action when the variable RUN_E2E is set
One (all?) of our contributors got annoying email once an hour on
their fork of Ubicloud.

To change the default, while allowing downstream forks to in principle
run E2E if they set everything up, I use a "variable" on the
repository.  A fork can set the variable at a URL like this:

    https://github.com/$user-name/ubicloud/settings/variables/actions

To enable the hourly E2E runs, assign the text "true" to the `RUN_E2E`
variable.