Fix nondeterministic test when using invalid flavor
This spec has failed twice on GitHub Actions when submitting PRs,
and once when I was doing parallel testing.

Reading through the code, the problem is Clover is redirecting based
on the Referer header, but I think Referer header state may be leaking
between test runs.  I'm not sure how to fix that, but this works
around the problem by setting a specific Referer header in the spec.