Replace database_cleaner-sequel with call to DB.transaction
I'm starting work on optimizing running single specs.  The TRUNCATE
commands run by database_cleaner-sequel take far longer than the
running a single spec.  Using a separate transaction per spec,
which is rolled back after the spec completes, is simple to do
by calling DB.transaction with the :rollback option.

With this, we can remove the GRANT TRUNCATE in the Rakefile,
as well as a substantial amount of setup code in spec/spec_config.

On my laptop, this cuts the time to run a single spec by about
45%, from about 2.5 seconds to about 1.4 seconds (this will
hopefully get to about 0.5 seconds with some additional changes).

Unfortunately, rspec does not support a global around(:context),
so you cannot do a transaction-per-suite and savepoint-per-spec
like you can when using minitest with minitest-hooks.