Update openapi.yml with closer-to-correct in-progress version
Daniel Farina <daniel@ubicloud.com> writing the commit message here.

Since the first prototypes of `openapi.yml`, there has been drift in
the implementation while we work out the development tools to run the
tests with the `committee` gem to keep it up to date.

But, as we're working with documentation, it'd be useful to have a
recent snapshot of the in-progress work to review how it looks when
generating API reference pages, so, copy in `geemus
<geemus@gmail.com>`'s up to date copy of the converging `openapi.yml`.
Since it's his in-progress work, I have listed him as the author.