Update data plane Postgres scripts to accept version as parameter
We used to hardcode 16 as the version, mostly while generating paths. With the
PG17 support, those paths changed to contain 17 instead of 16. Data plane does
not know the installed PostgreSQL version, so it needs that information from
control plane. With this commit we are updating data plane scripts to accept
version parameter from control plane.