Add PG17 base images
As first step of PG17 support, we generated base images that comes with PG17
and extensions compiled for PG17. We generate base images for standard and
ParadeDB flavors. Lantern flavor does not support PG17 yet. Upcoming commit
will ensure that we wouldn't try to provision a Lantern instance with PG17.