Add LB custom hostname feature to API
The client can pass a hostname that has a dns zone setup in ubicloud
in their project. We make sure that the project access is properly set
in the API and we re-check the project level access in the assemble.
This will be pretty useful for customers that have full control of their
hostname and we can generate DNS names, certificates for their apps in
ubicloud.