Fix NoMethodError when logging in user via remember token
When a user is logged in via the remember token, @current_user was
previously set to nil, because it was set before load_memory,
probably to make it available in the rodauth routes (I'm guessing).

If @current_user isn't already set after requiring authentication,
then set it at the point so it is available in all of the routes.