Add dns ipv4 to the dhcp payload and remove static dns conf
This is necessary because if we don't add the ipv4 address to the dhcp
payload, systemd-resolved kind of resolves it from the interface it gets
the dhcp packs from and adds the first address from the subnet as the
dns ipv4 address because the first address is added to the tap device.

We also remove the static dns configuration because that's unnecessary,
the dns addresses must be configured as part of dhcp.