Add ruby-lsp to Gemfile
Will Leinweber told me I was missing out if I wasn't using this.  I
gave it a try, unless my previous (few years ago) experiences with lsp
servers in emacs for ruby, this one worked immediately, having tested
finding declarations and references for a couple of symbols.

So, make it easier for everyone by adding it.