Log web exceptions with Clog
When we encounter unexpected errors in our web application, we simply
print the stack trace to stderr. This makes it challenging to
investigate the error cause, as our log service can only parse JSON
logs. Thus, we can log the exceptions with Clog, similar to how we
handle exceptions in the respirate.